+++ test/tutorial-2.dart
@@ -40,7 +40,7 @@ Future<Response> _handleTopLevel(Request req) async {
 }
 
 Future<Response> _myExceptionHandler(
-    Request req, Object ex, StackTrace? st) async {
+    Request req, Object ex, StackTrace st) async {
   int status;
   String message;