+++ test/tutorial-3.dart
@@ -34,7 +34,7 @@ Future<Response> _handleTopLevel(Request req) async {
   name = (name.isEmpty) ? 'world' : name;
 
   final resp = ResponseBuffered(ContentType.html)..write('''
-<html>
+<html lang="en">
   <head>
     <title>Woomera Tutorial</title>
   </head>
@@ -48,7 +48,7 @@ Future<Response> _handleTopLevel(Request req) async {
 
 Future<Response> _handleTestForm(Request req) async {
   final resp = ResponseBuffered(ContentType.html)..write('''
-<html>
+<html lang="en">
   <head>
     <title>Woomera Tutorial</title>
   </head>
@@ -69,7 +69,7 @@ Future<Response> _handleTestForm(Request req) async {
 }
 
 Future<Response> _myExceptionHandler(
-    Request req, Object ex, StackTrace? st) async {
+    Request req, Object ex, StackTrace st) async {
   int status;
   String message;
 
@@ -87,7 +87,7 @@ Future<Response> _myExceptionHandler(
   final resp = ResponseBuffered(ContentType.html)
     ..status = status
     ..write('''
-<html>
+<html lang="en">
   <head>
     <title>Error</title>
   </head>