+++ test/session_test.dart
@@ -10,9 +10,8 @@ import 'dart:async';
 import 'dart:convert' show utf8;
 import 'dart:io' show ContentType, HttpClient, HttpStatus;
 
-import 'package:test/test.dart';
 import 'package:logging/logging.dart';
-
+import 'package:test/test.dart';
 import 'package:woomera/woomera.dart';
 
 //================================================================